added main container to pass the props down to treeGraph
